11025 - Renamed MapVisualization and moved to container folder
